# docs/swagger/codebase-backend.yaml
openapi: 3.0.0
info:
  title: codebase-service-user
  description: This collection provide all endpoints of backend codebase which implement user service
  contact: {}
  version: '1.0'
servers:
  - url: '{protocol}://{host}:{port}/{service}'
    variables:
      protocol:
        enum:
          - http
          - https
        default: http
      host:
        default: localhost
      port:
        default: '9000'
      service:
        default: users
paths:
...
  /v1/refresh-token:
    post:
      tags:
        - v1
      summary: Refresh Token
      description: |-
        API endpoint to generate a new pair of token

        > A successful request returns new access token and refresh token
      operationId: RefreshToken
      parameters: []
      requestBody:
        content:
          application/x-www-form-urlencoded:
            encoding: {}
            schema:
              required:
                - refreshToken
              type: object
              properties:
                refreshToken:
                  type: string
                  description: Current refresh token
        required: false
      responses:
        '200':
          description: OK
          headers: {}
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RefreshTokenSuccess'
        '403':
          description: Forbidden
          headers: {}
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InvalidRefreshToken'
      deprecated: false
...
  /v2/verify:
    post:
      tags:
        - v2
      summary: Verify OTP
      description: |-
        API endpoint to Verify a registered user

        > A successful request returns success message
      operationId: VerifyOTP
      parameters: []
      requestBody:
        content:
          application/x-www-form-urlencoded:
            encoding: {}
            schema:
              required:
                - username
                - otp
              type: object
              properties:
                username:
                  type: string
                  description: Username which registered in the system
                otp:
                  type: string
                  description: OTP send to user by email or sms contain 6 characters of number
        required: false
      responses:
        '200':
          description: OK
          headers: {}
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/VerifyOTPSuccessV2'
        '404':
          description: Not Found
          headers: {}
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OTPNotFound'
        '409':
          description: Conflict
          headers: {}
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OTPDoesNotmatch'
      deprecated: false
...
components:
  schemas:
...
    ResponseDataLoginV1:
      title: ResponseDataLoginV1
      required:
        - token
        - refreshToken
      type: object
      properties:
        token:
          type: string
        refreshToken:
          type: string
      example:
        token: eyJhbGciOiJSUzI1NiIsInR5cCI6IkpXVCJ9.eyJ1c2VybmFtZSI6ImZyaXNraWdhdHJhIiwic3ViIjoiNjFiODI2ODBkMmE1MDIwODYzNmM4MTIwIiwiaWF0IjoxNjM5NDU4NzQ4LCJleHAiOjE2Mzk0NTk2NDgsImF1ZCI6Ijk3YjMzMTkzLTQzZmYtNGU1OC05MTI0LWIzYTliOWY3MmMzNCIsImlzcyI6InRlbGtvbWRldiJ9.p_aD_arc_LIPzomXE7NboTrossbqZ_he4q-YlGRJCf857zgoaaWLNgpEWrPb15lgC1y-H4CXhI2ATGVuGOs73EP_xVKwucApeon1cuVP2Ui4O2w3nZ0NU12i7k9aQogJzLhlObuQzt8Amigbuvo_tM6PYN7dRKpcsFxpkjHmnM_9It2XWTT3UDdZcnrl2DSdNm3kaBus0xJWMkwdM8UsqMZSWSaA8rKlPYhiKZWWRDCtIPIsJfAQDMRvIGzUIu1pcmrsNPdA3CKoBr2bWKDD_kKNwoLVxH43T9sUVMMW3rLd8BdZTkNDZH7Ym9NJ7HTN_A-q5iNdDmN7yZWG10P6k8VvgV6TNYU3UnhcVgUwlhNr9ICqYBQkS0W0e2ijwCte6nLGi9sBc1CmcnB66cOLoiR1_gyq_SxTMRyzsx2lbD5Vc-sXbvaHfTjpE2k-CK0Aixr4gVVV2_lVtLNsx7enPzLpN6HFjmK8qIGbjxVjjVovhXW0ihIYHskUvJ1cf0OU9aQn4TV-tRki6gj6HbN6vOKM1xoiqLO4u-EbGf-TXdooZf09Ay8S9UUsxeMJJRQAut1WX_x79b68k5HVC8L0aOe-H9HH5uEwY97_tMQwCBM8wAOp8VniOolDPxG6t7L_T_BHwmVGF1h-viTMBGwI5MWeL5UAzugOpi-9xrxeJzc
        refreshToken: eyJhbGciOiJSUzI1NiIsInR5cCI6IkpXVCJ9.eyJ1c2VybmFtZSI6ImZyaXNraWdhdHJhIiwic3ViIjoiNjFiODI2ODBkMmE1MDIwODYzNmM4MTIwIiwiaWF0IjoxNjM5NDU4NzQ4LCJleHAiOjE2NDAwNjM1NDgsImF1ZCI6Ijk3YjMzMTkzLTQzZmYtNGU1OC05MTI0LWIzYTliOWY3MmMzNCIsImlzcyI6InRlbGtvbWRldiJ9.tgRpTwfSh1vVeKRx-PNVn5dJHziXrLqqdIAOsALuie67zMy1R92sWpRMci1e-tslKS0KLvygoMOk77_PsBFhf4hJ1A1oWByyBTNn7N2YJApoAl-U03nsN8bZY35k1CNG3zh12YPzKOzh1sBsbOgoa3L6m669hWxdMDzPjabgJx_hHnGZ7ezn7AjiWwI34yAyEREvQYxdje9WzpD54uO1fIYcdnC_TofnUy2lp634Kk-7IPEbqrO6ZqR7j6LY-EjMUkzWddXAr7lED-kfHWUhroWa6U6lk-NTdM0c6g1mol4G8_RycfJ5K9l3gl_9WxnUHceGE7sRtP03_tmyRWEyKodoPFYp6z5Mnb89l6-Lg3XnOPz3tl6wr_inK2vO7BD3VfSZhIdf5Nf-wTABuzajzeB3ssUgYXSFCO6rtu53i58LWIFBSnVfI_jpEMOcpvkEoDjGhWQ9WW-_nfakjv8vAQr_7yZertx4wsWwWx38pwXKsnE9YB6UXdzIRVMo-eCQrlah_47hWJbBaSyZKDoM4D8B6clShH9xsUYt7h2BH6JVm-TtlUhI4VNZvINoM1oAKeYtZ9IFIVoLc8h2QpZam7KwLeU59ZN8RNHoXgXqTTDci5WSSs4y0bs3mMQnlxvfwADW_iDKb836JWL9E6v0hpyZStCpYs0MLumLvz4CVjk
...
    RefreshTokenSuccess:
      title: RefreshTokenSuccess
      required:
        - success
        - data
        - message
        - code
      type: object
      properties:
        success:
          type: boolean
        data:
          $ref: '#/components/schemas/ResponseDataLoginV1'
        message:
          type: string
        code:
          type: integer
          format: int32
      example:
        success: true
        data:
          token: eyJhbGciOiJSUzI1NiIsInR5cCI6IkpXVCJ9.eyJ1c2VySWQiOiI2MWI4MjY4MGQyYTUwMjA4NjM2YzgxMjAiLCJpYXQiOjE2Mzk0NTkyNzMsImV4cCI6MTYzOTQ2MDE3MywiYXVkIjoiOTdiMzMxOTMtNDNmZi00ZTU4LTkxMjQtYjNhOWI5ZjcyYzM0IiwiaXNzIjoidGVsa29tZGV2In0.CDVeE3ZCJJKAzv_yCHz3lYcjPuKL40GdH92fi3m39dxseM49SCGFeyx_jrFwaCTy0Y2IvyF_HqphKPrHo8prjvUxU6Q1bAJObxmX_Eq4Kt0Nm4ZKXMWZQdlkfn6-xj3PkrchQ-DdI_oDux5BeNm4JXkckE2kZCSd-WCwyTgTL7NZvs6o8Zk-kZCe4QUwiyye6tu4Nsn_vequ22-nkbmZiMtYicNyRzVLsqJsSAmU06tQFIr4hMxajsWFkJmHbpaS8Sm20PXCsDKsiqQGC2rFYWefsGloT3nozPLweNqGFcFHu5AZlYGueh1xBAgiURlrLMEAbIsq-WiQlA_97bNkR92wz3PDKGB_UCwEMGIZ_1CL8Dp3bThdjo9-bvA-bXOknRKcEY0rZC_s6s9rCHK94jjBuRyIGZJCfkTW4zFV6g7qJ_iosL7qHuDN-RWSYwymMxdc6INtRkypia09bx9-CBnRgoEqeDZ2HNprfK9eKB2O_JYApPgwfATNObAkH8uABh8oXtJmc4M-FeQCGjO43Rs_qoXbwRgKscLh-AWa5IOzNCwSVS72wWwocnhVR4zladyRUdQeqTcTxxSlYkDRSmy7yJI_qdjUQO5axe5XaYgFYlkQpwIxCQL0M8CefHCLRtLJAPA4MW4rIchm_S_5XrLXqfrdxXY4fsawhnoG0yk
          refreshToken: eyJhbGciOiJSUzI1NiIsInR5cCI6IkpXVCJ9.eyJ1c2VySWQiOiI2MWI4MjY4MGQyYTUwMjA4NjM2YzgxMjAiLCJpYXQiOjE2Mzk0NTkyNzMsImV4cCI6MTY0MDA2NDA3MywiYXVkIjoiOTdiMzMxOTMtNDNmZi00ZTU4LTkxMjQtYjNhOWI5ZjcyYzM0IiwiaXNzIjoidGVsa29tZGV2In0.rlBbpk9RO2oWUgaoWH668h6-H20T6I-jhPXPfegBJkz755Eg1vqSEBSSY6OnzH44PjBjJPJRHkr8_LSu2TLbidjHYblCKVEH4qdWTHIWif-zID2O5YT_M5FpO5zV_sPFxQWDnbgLo5Tumecfhu80Ewe3gL6oI5aIzdd7YNd6drC1PXgU6Rcs9Fv8TISgJyOCyd_R7TTR9LaOtIlDut-Jh3TzzncWCKJM0p7pVEuGkfK0j9yj6Lb7P9VbEX6AJFlQyIltnPfxHSLVXc4uCheHQ_rbriNggUQ4aNxZmBThi7Sywy1A-208ZMMcR7IbaDd9IGz0E3slBWJzPpI4IHzGc1OyLbvCU3ldd8j0u9kEYnvwMhqXRsWw_w8TXGr0dDTyXlPj8GuzqIS2h28a1JioAqBwB03z8lbDbOxaaJOKu6VNFg2ehAUIDONkpEOLS5GVpZX43z6gtJhcvG7lQcNsKo4hEQ9-ZnivP7_n5Jh3DlgJNK5pg-MpTR6IT96MOOjKVrArxxnmWYYo6eHCTfsIDFuf9VtXmOVNqqwS3y77icv_JMC-yMK27FANlXTT41eFLmVowGAZJwNd-yhVWDzkRIVFMnS1IgeQjZafeoGYla_NcwZIOyOwUhRmrPB0X5-JfcckYYAGE8XcvGYSig79iNClxygPvnFRnrlVGEizmYQ
        message: Get Refresh Token
        code: 200
    InvalidRefreshToken:
      title: InvalidRefreshToken
      required:
        - success
        - data
        - message
        - code
      type: object
      properties:
        success:
          type: boolean
        data:
          type: string
        message:
          type: string
        code:
          type: integer
          format: int32
      example:
        success: false
        data: ''
        message: Token is not valid
        code: 403
...
    VerifyOTPSuccessV2:
      title: VerifyOTPSuccessV2
      required:
        - success
        - data
        - message
        - code
      type: object
      properties:
        success:
          type: boolean
        data:
          type: string
        message:
          type: string
        code:
          type: integer
          format: int32
      example:
        success: true
        data: success
        message: Logout User
        code: 200
    OTPNotFound:
      title: OTPNotFound
      required:
        - success
        - data
        - message
        - code
      type: object
      properties:
        success:
          type: boolean
        data:
          type: string
        message:
          type: string
        code:
          type: integer
          format: int32
      example:
        success: false
        data: ''
        message: user's otp not found
        code: 404
    OTPDoesNotmatch:
      title: OTPDoesNotmatch
      required:
        - success
        - data
        - message
        - code
      type: object
      properties:
        success:
          type: boolean
        data:
          type: string
        message:
          type: string
        code:
          type: integer
          format: int32
      example:
        success: false
        data: ''
        message: user's otp doesn't match
        code: 409
...
  securitySchemes:
    basicAuth:
      type: http
      scheme: basic
    bearerAuth:
      type: http
      scheme: bearer
security:
  - basicAuth: []
  - bearerAuth: []
...